JIRA LSC-2105, MMS ID, EDI
https://jira.library.ucla.edu/browse/LSC-2105
Added a scope and contents to box 56 to say: This box contains photographs that were used in Hilda Kuper's book, <title>The Swazi</title>.
Updated the catalog record ID in General Note. Added ethical description statement as well as updated the Processing info note.
Took out ?s in the descriptions as well as removed ns2 tags.